+++ CHANGELOG.md
@@ -6,6 +6,17 @@ Most recent releases are shown at the top. Each release shows:
 - **Changed**: Additional parameters, changes to inputs or outputs, etc
 - **Fixed**: Bug fixes that don't change documented behaviour
 
+## 0.21.8 (2026-03-23)
+
+### new:
+- N/A
+
+### changed:
+- N/A
+
+### fixed:
+- Include AWS standard names for Bedrock endpoint URL
+
 ## 0.21.7 (2026-03-20)
 
 ### new: